20Erfassen/requirements.txt
mysql-connector-python
#delorean
argparse
black